Fix cross-build
