+++ pkg/kubelet/cadvisor/BUILD
@@ -11,11 +11,13 @@ go_library(
     srcs = [
         "cadvisor_unsupported.go",
         "doc.go",
+        "helpers_unsupported.go",
         "types.go",
         "util.go",
     ] + select({
         "@io_bazel_rules_go//go/platform:linux_amd64": [
             "cadvisor_linux.go",
+            "helpers_linux.go",
         ],
         "@io_bazel_rules_go//go/platform:windows_amd64": [
             "cadvisor_windows.go",
@@ -26,7 +28,6 @@ go_library(
         "//pkg/api/v1/helper:go_default_library",
         "//pkg/features:go_default_library",
         "//vendor/github.com/google/cadvisor/events:go_default_library",
-        "//vendor/github.com/google/cadvisor/fs:go_default_library",
         "//vendor/github.com/google/cadvisor/info/v1:go_default_library",
         "//vendor/github.com/google/cadvisor/info/v2:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
@@ -38,6 +39,7 @@ go_library(
             "//vendor/github.com/golang/glog:go_default_library",
             "//vendor/github.com/google/cadvisor/cache/memory:go_default_library",
             "//vendor/github.com/google/cadvisor/container:go_default_library",
+            "//vendor/github.com/google/cadvisor/fs:go_default_library",
             "//vendor/github.com/google/cadvisor/http:go_default_library",
             "//vendor/github.com/google/cadvisor/manager:go_default_library",
             "//vendor/github.com/google/cadvisor/metrics:go_default_library",
