+++ pkg/kubelet/cadvisor/cadvisor_windows.go
@@ -77,3 +77,7 @@ func (cu *cadvisorClient) WatchEvents(request *events.Request) (*events.EventCha
 func (cu *cadvisorClient) HasDedicatedImageFs() (bool, error) {
 	return false, nil
 }
+
+func (c *cadvisorClient) GetFsInfoByFsUUID(uuid string) (cadvisorapiv2.FsInfo, error) {
+	return cadvisorapiv2.FsInfo{}, nil
+}
